Update gripper control algorithm
This version extends the rotation range from -90 to 90 degrees to -180 to 180 degrees. The motor angle provided by this algorithm can be used to control the gripper hardware. This would be the foundation for the next stage.

The next step is to test the algorithm on the hardware, once the function is confirmed working, a feedback loop that minimises the error that appears in the hardware control will be needed.

From the algorithm perspective, the next step is to add an optimisation algorithm that minimises the angle needed to complete the movement.